+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewPage.js
@@ -83,7 +83,7 @@ export const RunsPerDataPassOverviewPage = ({
         globalAggregatedQuality: {
             name: 'GAQ',
             visible: true,
-            format: (_, { runNumber }) => frontLink('GAQ', 'gaq-flags', { dataPassId, runNumber }),
+            format: (_, { runNumber }) => frontLink(h('button.btn.btn-primary', 'GAQ'), 'gaq-flags', { dataPassId, runNumber }),
             profiles: ['runsPerDataPass'],
         },
         ...createRunDetectorsAsyncQcActiveColumns(
